requirements.txt
amqp==1.4.9
anyjson==0.3.3
astroid==1.4.8
Babel==2.3.4
backports.functools-lru-cache==1.2.1
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.23
bleach==1.4.3
celery==3.1.23
certifi==2016.2.28
configparser==3.5.0
Django==1.9.7
django-allauth==0.25.2
django-bootstrap-pagination==1.6.1
django-forms-bootstrap==3.0.1
django-guardian==1.4.5
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-markdownx==1.6.3
django-phonenumber-field==1.1.0
django-redis-cache==1.6.5
django-versatileimagefield==1.5
django-websocket-redis==0.4.6
flower==0.9.1
futures==3.0.5
geoip2==2.4.0
gevent==1.1.2
greenlet==0.4.10
html5lib==0.9999999
ipaddress==1.0.17
isort==4.2.5
jsonfield==1.0.3
kombu==3.0.35
lazy-object-proxy==1.2.2
Markdown==2.6.6
markdown2==2.3.1
maxminddb==1.2.1
mccabe==0.5.2
oauthlib==2.0.0
phonenumberslite==7.4.4
Pillow==3.3.1
psycopg2==2.6.2
PyJWT==1.4.0
python-http-client==1.2.3
python-openid==2.2.5
pytz==2016.4
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.7.0
sendgrid==2.2.1
sendgrid-django==2.0.0
six==1.10.0
smtpapi==0.3.1
tornado==4.2
tweepy==3.5.0
wrapt==1.10.8
